Fix init of best loss
If --use-best-param is enabled and optimizer is not spsa best_loss should be 1-best_result_threshold. The display of best loss is fixed. The optimizer is not affected because if the init loss is 0.55, it will attempt to reduce it, if init loss is 0.45 it will also attempt to reduce it.